# src/Ims.Case.Api/wwwroot/docs/case-v1p0-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Competencies and Academic Standards Exchange (CASE) Service OpenAPI (JSON) Definition",
    "description": "The Competencies and Academic Standards Exchange (CASE) Service enables the exchange of data between a Competency Records Service Provider and the consumers of the associated data. This service has been described using the IMS Model Driven Specification development approach this being the Platform Specific Model (PSM) of the service.",
    "termsOfService": "http://www.opendefinition.org/licenses/odc-by",
    "contact": {
      "name": "Brad R. Marshall ",
      "email": "bmarshall@sped.mobi"
    },
    "license": {
      "name": "Open Data Commons Attribution License (ODCAL)",
      "url": "http://www.opendefinition.org/licenses/odc-by"
    },
    "version": "v1p0"
  },
  "servers": [
    {
      "url": "https://teks-api.texasgateway.org/ims/case/v1p0",
      "description": "TEA Texas Gateway Server"
    }
  ],
...
  "paths": {
...
    "/CFDocuments": {
      "get": {
        "tags": [
          "DocumentsManager"
        ],
        "summary": "The REST read request message for the getAllCFDocuments() API call.",
        "description": "This is a request to the Service Provider to provide all of the Competency Framework Documents. ",
        "operationId": "getAllCFDocuments",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "This is used as part of the data pagination mechanism to control the download rate of data. The 'limit' defines the download segmentation value i.e. the maximum number of records to be contained in the response. The form of implementation is described in the corresponding binding document(s).",
            "required": false,
            "allowEmptyValue": false,
            "style": "form",
            "explode": true,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "This is used as part of the data pagination mechanism to control the download rate of data. The 'offset' is the number of the first record to be supplied in the segmented response message. The form of implementation is described in the corresponding binding document(s).",
            "required": false,
            "allowEmptyValue": false,
            "style": "form",
            "explode": true,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "This is used as part of the sorting mechanism to be use by the service provider. The 'sort' identifies the sort criteria to be used for the records in the response message. Use with the orderBy parameter. The form of implementation is described in the corresponding binding document(s).",
            "required": false,
            "allowEmptyValue": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "This is used as part of the sorting mechanism to be use by the service provider. This defines the form of ordering for response to the sorted request i.e. ascending (asc) or descending (desc). The form of implementation is described in the corresponding binding document(s).",
            "required": false,
            "allowEmptyValue": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "filter",
            "in": "query",
            "description": "This is used for the data filtering mechanism to be applied by the service provider. It defines the filtering rules to be applied when identifying the records to be supplied in the response message. The form of implementation is described in the corresponding binding document(s).",
            "required": false,
            "allowEmptyValue": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "This is used as part of the field selection mechanism to be applied by the service provider. This identifies the range of fields that should be supplied in the response message. The form of implementation is described in the corresponding binding document(s).",
            "required": false,
            "allowEmptyValue": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "This is the response when the request has been completed successfully. It is the set of CFDocuments from the service provider.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CFDocumentSet"
                }
              }
            }
          },
          "400": {
            "description": "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "403": {
            "description": "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "429": {
            "description": "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "500": {
            "description": "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "default": {
            "description": "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          }
        }
      }
    },
    "/CFDocuments/{sourcedId}": {
      "get": {
        "tags": [
          "DocumentsManager"
        ],
        "summary": "The REST read request message for the getCFDocument() API call.",
        "description": "This is a request to the service provider to provide the information for the specific Competency Framework Document. If the identified record cannot be found then the 'unknownobject' status code must be reported.",
        "operationId": "getCFDocument",
        "parameters": [
          {
            "name": "sourcedId",
            "in": "path",
            "description": "The UUID that identifies the Competency Framework Document that is to be read from the service provider.",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "This is the response when the request has been completed successfully. It is the CFDocument from the service provider.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CFDocument"
                }
              }
            }
          },
          "400": {
            "description": "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "403": {
            "description": "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "429": {
            "description": "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "500": {
            "description": "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "default": {
            "description": "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CFDocument": {
        "type": "object",
        "description": "The container for the data about a competency framework document (CFDocument) when exchanged outside of the context of a CFPackage. A CFDocument is the root for the creation of a learning standard/competency.",
        "allOf": [
          {
            "$ref": "#/components/schemas/CFPckgDocument"
          },
          {
            "required": [
              "CFPackageURI"
            ],
            "type": "object",
            "properties": {
              "CFPackageURI": {
                "$ref": "#/components/schemas/LinkURI"
              }
            },
            "additionalProperties": false
          }
        ]
      },
      "CFDocumentSet": {
        "required": [
          "CFDocuments"
        ],
        "type": "object",
        "properties": {
          "CFDocuments": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CFDocument"
            }
          }
        },
        "additionalProperties": false,
        "description": "This is the container for a collection of CFDocuments. There must be at least one CFDocument."
      },
...
      "CFPckgDocument": {
        "required": [
          "creator",
          "identifier",
          "lastChangeDateTime",
          "title",
          "uri"
        ],
        "type": "object",
        "properties": {
          "identifier": {
            "$ref": "#/components/schemas/UUID"
          },
          "uri": {
            "type": "string",
            "description": "Model Primitive Datatype = AnyURI",
            "format": "uri"
          },
          "creator": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "title": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "lastChangeDateTime": {
            "type": "string",
            "description": "Model Primitive Datatype = DateTime",
            "format": "date-time"
          },
          "officialSourceURL": {
            "$ref": "#/components/schemas/URL"
          },
          "publisher": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "description": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "subject": {
            "minItems": 0,
            "type": "array",
            "description": "Model Primitive Datatype = NormalizedString",
            "items": {
              "type": "string"
            }
          },
          "subjectURI": {
            "minItems": 0,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinkURI"
            }
          },
          "language": {
            "type": "string",
            "description": "Model Primitive Datatype = Language"
          },
          "version": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "adoptionStatus": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "statusStartDate": {
            "type": "string",
            "description": "Model Primitive Datatype = Date",
            "format": "date"
          },
          "statusEndDate": {
            "type": "string",
            "description": "Model Primitive Datatype = Date",
            "format": "date"
          },
          "licenseURI": {
            "$ref": "#/components/schemas/LinkURI"
          },
          "notes": {
            "type": "string",
            "description": "Model Primitive Datatype = String"
          }
        },
        "additionalProperties": false,
        "description": "The container for the data about a competency framework document (CFDocument) within a CFPackage. A CFDocument is the root for the creation of a learning standard/competency."
      },
...
      "LinkURI": {
        "required": [
          "identifier",
          "title",
          "uri"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "identifier": {
            "$ref": "#/components/schemas/UUID"
          },
          "uri": {
            "type": "string",
            "description": "Model Primitive Datatype = AnyURI",
            "format": "uri"
          }
        },
        "additionalProperties": false,
        "description": "A container for the information that is used to achieve the link data reference."
      },
      "URL": {
        "type": "string",
        "description": "The data-type for establishing a Uniform Resource Locator (URL) as defined by W3C.",
        "format": "uri"
      },
      "UUID": {
        "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[1-5]{1}[0-9a-f]{3}-[8-9a-b]{1}[0-9a-f]{3}-[0-9a-f]{12}",
        "type": "string",
        "description": "The data-type for establishing a Globally Unique Identifier (GUID). The form of the GUID is a Universally Unique Identifier (UUID) of 16 hexadecimal characters (lower case) in the format 8-4-4-4-12. All permitted versions (1-5) and variants (1-2) are supported.",
        "format": "guid"
      },
      "imsx_CodeMinor": {
        "required": [
          "imsx_codeMinorField"
        ],
        "type": "object",
        "properties": {
          "imsx_codeMinorField": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/imsx_CodeMinorField"
            }
          }
        },
        "additionalProperties": false,
        "description": "This is the container for the set of code minor status codes reported in the responses from the Service Provider."
      },
      "imsx_CodeMinorField": {
        "required": [
          "imsx_codeMinorFieldName",
          "imsx_codeMinorFieldValue"
        ],
        "type": "object",
        "properties": {
          "imsx_codeMinorFieldName": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "imsx_codeMinorFieldValue": {
            "type": "string",
            "enum": [
              "fullsuccess",
              "invalid_sort_field",
              "invalid_selection_field",
              "forbidden",
              "unauthorisedrequest",
              "internal_server_error",
              "unknownobject",
              "server_busy",
              "invaliduuid"
            ]
          }
        },
        "additionalProperties": false,
        "description": "This is the container for a single code minor status code."
      },
      "imsx_StatusInfo": {
        "required": [
          "imsx_codeMajor",
          "imsx_severity"
        ],
        "type": "object",
        "properties": {
          "imsx_codeMajor": {
            "type": "string",
            "enum": [
              "success",
              "processing",
              "failure",
              "unsupported"
            ]
          },
          "imsx_severity": {
            "type": "string",
            "enum": [
              "status",
              "warning",
              "error"
            ]
          },
          "imsx_description": {
            "type": "string",
            "description": "Model Primitive Datatype = String"
          },
          "imsx_codeMinor": {
            "$ref": "#/components/schemas/imsx_CodeMinor"
          }
        },
        "additionalProperties": false,
        "description": "This is the container for the status code and associated information returned within the HTTP messages received from the Service Provider. For the CASE service this object will only be returned to provide information about a failed request i.e. it will NOT be in the payload for a successful request. See Appendix B for further information on the interpretation of the information contained within this class"
      }
    }
  }
}